refactor(Answer): 도메인의 유효성 검증 분리

+++ src/main/java/qna/domain/Answer.java
@@ -1,5 +1,6 @@
 package qna.domain;
 
+import qna.CannotDeleteException;
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
 
@@ -68,9 +69,16 @@ public class Answer extends AbstractEntity {
         this.question = question;
     }
 
-    public DeleteHistory delete() {
+    public DeleteHistory delete(User loginUser) throws CannotDeleteException {
+        validateAuthority(loginUser);
         this.setDeleted(true);
         return new DeleteHistory(ContentType.ANSWER, getId(), getWriter());
+    }
+
+    private void validateAuthority(User loginUser) throws CannotDeleteException {
+        if (!this.isOwner(loginUser)) {
+            throw new CannotDeleteException("다른 사람이 쓴 답변이 있어 삭제할 수 없습니다.");
+        }
     }
 
     @Override
